+++ packages/astro/test/astro-sync.test.ts
@@ -16,7 +16,7 @@ const createFixture = () => {
 
 	return {
 		async load(root: string) {
-			astroFixture = await loadFixture({ root });
+			astroFixture = await loadFixture({ root, outDir: './dist/astro-sync-1/' });
 			return astroFixture.config;
 		},
 		clean() {
@@ -255,7 +255,10 @@ describe('astro sync', () => {
 				},
 			});
 
-			const astroFixture = await loadFixture({ root: './fixtures/astro-basic/' });
+			const astroFixture = await loadFixture({
+				root: './fixtures/astro-basic/',
+				outDir: './dist/astro-sync-no-content-config/',
+			});
 			fs.rmSync(new URL('./.astro/', astroFixture.config.root), { force: true, recursive: true });
 
 			// @ts-ignore