+++ packages/astro/test/chrome-devtools-workspace.test.ts
@@ -13,6 +13,7 @@ describe('Chrome DevTools workspace', () => {
 				experimental: {
 					chromeDevtoolsWorkspace: true,
 				},
+				outDir: './dist/chrome-devtools-workspace-with-experimental-flag-enabled/',
 			});
 			devServer = await fixture.startDevServer();
 		});
@@ -45,6 +46,7 @@ describe('Chrome DevTools workspace', () => {
 				experimental: {
 					chromeDevtoolsWorkspace: false,
 				},
+				outDir: './dist/chrome-devtools-workspace-with-experimental-flag-disabled/',
 			});
 			devServer = await fixture.startDevServer();
 		});