+++ packages/astro/test/html-slots.test.ts
@@ -9,6 +9,7 @@ describe('HTML Slots', () => {
 	before(async () => {
 		fixture = await loadFixture({
 			root: './fixtures/html-slots/',
+			outDir: './dist/html-slots/',
 		});
 	});