+++ packages/astro/test/non-html-pages.test.ts
@@ -6,7 +6,10 @@ describe('Non-HTML Pages', () => {
 	let fixture: Fixture;
 
 	before(async () => {
-		fixture = await loadFixture({ root: './fixtures/non-html-pages/' });
+		fixture = await loadFixture({
+			root: './fixtures/non-html-pages/',
+			outDir: './dist/non-html-pages/',
+		});
 		await fixture.build();
 	});