+++ packages/astro/test/sessions.test.ts
@@ -19,6 +19,7 @@ describe('Astro.session', () => {
 					driver: 'fs',
 					ttl: 20,
 				},
+				outDir: './dist/sessions-production/',
 			});
 			await fixture.build({});
 			app = await fixture.loadTestAdapterApp();
@@ -223,6 +224,7 @@ describe('Astro.session', () => {
 					driver: 'fs',
 					ttl: 20,
 				},
+				outDir: './dist/sessions-development/',
 			});
 			devServer = await fixture.startDevServer();
 		});