+++ packages/integrations/markdoc/components/TreeNode.ts
@@ -88,6 +88,7 @@ function renderTreeNodeToFactoryResult(result: SSRResult, treeNode: TreeNode) {
 		const head = unescapeHTML(styles + links + scripts);
 
 		let headAndContent = createHeadAndContent(
+			// @ts-expect-error: `createHeadAndContent` expect a string and we know `head` is definitely a string
 			head,
 			renderTemplate`${renderComponent(
 				result,
@@ -112,6 +113,7 @@ function renderTreeNodeToFactoryResult(result: SSRResult, treeNode: TreeNode) {
 }
 
 export const ComponentNode = createComponent({
+	// @ts-expect-error: The types are a bit complex here. Some further refactor might be needed to make this more type-safe.
 	factory(result: SSRResult, { treeNode }: { treeNode: TreeNode | TreeNode[] }) {
 		return renderTreeNodeToFactoryResult(result, treeNode);
 	},