+++ packages/integrations/netlify/test/hosted/hosted-astro-project/src/middleware.ts
@@ -1,5 +1,6 @@
 import https from 'node:https';
 
+// @ts-expect-error: Parameter `context` implicitly has an any type
 export const onRequest = (context, next) => {
 	console.info(context.netlify);
 	context.locals.middleware = context?.locals?.netlify?.context?.geo?.country?.code ?? null;